requirements.txt
flask
pandas
numpy
scikit-learn
faiss-cpu
sentence-transformers
requests
beautifulsoup4
gunicorn

